+++ test/src/test_memory.cpp
@@ -7,13 +7,12 @@
 #include <gtest/gtest.h>
 
 #include <abacus/metrics.hpp>
-TEST(test_memory, default_constructor)
+TEST(test_memory, default_test)
 {
     uint64_t count = 0xFFFFFFFFFFFFFFFFU;
     uint16_t max_name_bytes = 32;
     uint16_t max_metrics = 32;
-    uint8_t level = 0xFF;
-    abacus::metrics metrics(max_metrics, max_name_bytes, "metrics", level);
+    abacus::metrics metrics(max_metrics, max_name_bytes, "metrics");
 
     for (std::size_t i = 0; i < max_metrics; i++)
     {