+++ NEWS.rst
@@ -6,7 +6,11 @@ every change, see the Git log.
 
 Latest
 ------
+* Major: Initial release
 * Major: Added level member to metrics class
 * Minor: Added read_storage() to metrics class
 * Major: Separated the metric class from metrics
+* Major: Added class view to handle the raw memory of class metrics
+* Minor: Added class view_iterator
+* Minor: Moved to_json() member function to detail